I know you think you won

+++ src/graphql/typeDefs.ts
@@ -44,8 +44,8 @@ const typeDefs = `
   }
 
   input Coordinate {
-    x: Int
-    y: Int
+    x: Float
+    y: Float
   }
 
   type Annotation {
